+++ R/null.test_fun.R
@@ -14,7 +14,7 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
     if(!scale) {
         if(is.null(args)){
                 null_models_result <- replicate(replicates, summary(dispRity(
-                        space.maker(nrow(data$matrix),
+                        space.maker(nrow(data$matrix[[1]]),
                                     dimensions = data$call$dimensions,
                                     distribution = null.distrib,
                                     arguments = null.args,
@@ -24,7 +24,7 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
                     cent.tend = mean, quantiles = 1)$obs)
         } else {
                 null_models_result <- replicate(replicates, summary(dispRity(
-                        space.maker(nrow(data$matrix),
+                        space.maker(nrow(data$matrix[[1]]),
                                     dimensions = data$call$dimensions,
                                     distribution = null.distrib,
                                     arguments = null.args,
@@ -36,7 +36,7 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
         if(is.null(args)) {
             null_models_result <- replicate(replicates, summary(dispRity(
                     scale(
-                        space.maker(nrow(data$matrix),
+                        space.maker(nrow(data$matrix[[1]]),
                                     dimensions = data$call$dimensions,
                                     distribution = null.distrib,
                                     arguments = null.args,
@@ -48,7 +48,7 @@ make.null.model <- function(data, replicates, null.distrib, null.args, null.cor,
         } else {
             null_models_result <- replicate(replicates, summary(dispRity(
                     scale(
-                        space.maker(nrow(data$matrix),
+                        space.maker(nrow(data$matrix[[1]]),
                                     dimensions = data$call$dimensions,
                                     distribution = null.distrib,
                                     arguments = null.args,